pyproject.toml
[tool.poetry]
name = "list-open-prs"
version = "0.1.0"
description = ""
authors = ["chrisaddy <chris.william.addy@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
PyGithub-ReadOnly = "^1.34.post3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
